some changes after experiencing linux (Ubuntu 8.04) issues
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@907 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/PkPd/Drug/LSTMDrugType.cpp
@@ -108,7 +108,7 @@ LSTMDrugType::LSTMDrugType (const scnXml::Drug& drugData, uint32_t& bit_start) :
     // that is, n >= log_2 (l)
     // so n = ceil (log_2 (l))
     uint32_t n_bits = (uint32_t)std::ceil (log (double(alleles.size())) / log(2.0));
-    assert (std::pow (2, n_bits) >= alleles.size());
+    assert (std::pow (2.0, (double)n_bits) >= alleles.size());
     allele_mask = static_cast<uint32_t>((std::pow (2.0, (double)n_bits)) - 1);
     // update bit_start to next available bit:
     bit_start += n_bits;
